+++ js/main.js
@@ -66,7 +66,7 @@ const HC_LIGHT = {
         }
     },
     ternaryAxis: {
-        plotOptions: {
+        common: {
             lineColor: '#d6d6d6',
             gridLineColor: '#d6d6d6',
             title: {
@@ -143,7 +143,7 @@ const HC_DARK = {
         }
     },
     ternaryAxis: {
-        plotOptions: {
+        common: {
             lineColor: '#30363D',
             gridLineColor: '#30363D',
             title: {